requirements.txt
alembic==1.0.5
boto3==1.9.71
botocore==1.12.71
Click==7.0
docutils==0.14
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
Mako==1.0.7
MarkupSafe==1.1.0
Pillow==5.3.0
PyMySQL==0.9.3
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
s3transfer==0.1.13
six==1.12.0
SQLAlchemy==1.2.15
urllib3==1.24.1
visitor==0.1.3
Wand==0.4.5
Werkzeug==0.14.1
WTForms==2.2.1
gunicorn==19.9.0
